+++ CHANGES.rst
@@ -4,6 +4,8 @@ Changelog
 6.2.25 (unreleased)
 -------------------
 
+- Add dependency with collective.volto.blocksfield >= 2.2.0 and install it to have all blocks indexed in block_types index.
+  [cekk]
 - Do not use eea.api.taxonomy because it is deprecated.
   [cekk]